refactor: use $lib
