--- src/data/posts.ts
@@ -1,11 +1,0 @@
-import type { CollectionEntry } from 'astro:content';
-
-export type LinkPost = CollectionEntry<'links'>['data'] & { type: 'LINK' };
-
-export type BlogPost = CollectionEntry<'blog'>['data'] & {
-    slug: string;
-    type: 'BLOG';
-};
-
-export type Post = LinkPost | BlogPost;
-
