+++ src/lib/components/LogoRainbow.svelte
