+++ src/lib/components/TRainbow.svelte
