+++ src/lib/components/posts/Posts.astro
@@ -1,6 +1,6 @@
 ---
 import { type Tag, tags } from '$data/tags';
-import { getPosts } from '$helpers/posts';
+import { getPosts } from '$lib/posts';
 import PostCard from './PostCard.astro';
 
 interface Props {
