+++ src/lib/components/posts/SettingsModal.svelte
@@ -56,7 +56,7 @@
 </Modal>
 
 <style lang="scss">
-    @import 'src/helpers/media';
+    @import 'src/lib/media';
 
     .input-group {
         width: 100%;
