+++ src/lib/posts.ts
@@ -1,5 +1,13 @@
-import { getCollection } from 'astro:content';
-import type { Post } from '../data/posts';
+import { getCollection, type CollectionEntry } from 'astro:content';
+
+export type LinkPost = CollectionEntry<'links'>['data'] & { type: 'LINK' };
+
+export type BlogPost = CollectionEntry<'blog'>['data'] & {
+    slug: string;
+    type: 'BLOG';
+};
+
+export type Post = LinkPost | BlogPost;
 
 export async function getPosts() {
     const linkPosts = await getCollection('links');
