+++ src/pages/404.astro
@@ -1,5 +1,5 @@
 ---
-import Global from '$layouts/Global.astro';
+import Global from '$lib/layouts/Global.astro';
 ---
 
 <Global>
