+++ src/pages/branding.astro
@@ -1,9 +1,9 @@
 ---
-import ColourGroup from '$components/branding/ColourGroup.astro';
-import Logo from '$components/Logo.svelte';
-import LogoRainbow from '$components/LogoRainbow.svelte';
+import ColourGroup from '$lib/components/branding/ColourGroup.astro';
+import Logo from '$lib/components/Logo.svelte';
+import LogoRainbow from '$lib/components/LogoRainbow.svelte';
 import { colours } from '$data/colours';
-import Page from '$layouts/Page.astro';
+import Page from '$lib/layouts/Page.astro';
 ---
 
 <Page>
