+++ src/pages/contact.astro
@@ -3,8 +3,8 @@ import { faDiscord, faTwitter } from '@fortawesome/free-brands-svg-icons';
 import { faMastodon } from '@fortawesome/free-brands-svg-icons';
 import { faEnvelope } from '@fortawesome/free-solid-svg-icons';
 import { faGithub } from '@fortawesome/free-brands-svg-icons';
+import Page from '$lib/layouts/Page.astro';
 import { Image } from 'astro:assets';
-import Page from '$layouts/Page.astro';
 import Fa from 'svelte-fa';
 ---
 
@@ -62,6 +62,7 @@ import Fa from 'svelte-fa';
         >
             <Image
                 width={75}
+                height={75}
                 src="https://cdn.discordapp.com/icons/755788441161302136/bc7d982c002e2786e34c17f58b848b06.webp"
                 alt="The Dev Lounge logo"
             />
@@ -78,7 +79,7 @@ import Fa from 'svelte-fa';
 </Page>
 
 <style lang="scss">
-    @import 'src/helpers/media';
+    @import 'src/lib/media';
 
     .devlounge {
         .card {
