+++ src/pages/donate.astro
@@ -1,8 +1,8 @@
 ---
 import { faPaypal, faPatreon } from '@fortawesome/free-brands-svg-icons';
 import { supporters, donators } from '$data/donators';
-import Donator from '$components/Donator.svelte';
-import Page from '$layouts/Page.astro';
+import Donator from '$lib/components/Donator.svelte';
+import Page from '$lib/layouts/Page.astro';
 import Fa from 'svelte-fa';
 ---
