+++ src/pages/index.astro
@@ -2,8 +2,8 @@
 import { faDiscord, faGithub } from '@fortawesome/free-brands-svg-icons';
 import { faMastodon } from '@fortawesome/free-brands-svg-icons';
 import { faTwitter } from '@fortawesome/free-brands-svg-icons';
-import Posts from '$components/posts/Posts.astro';
-import Page from '$layouts/Page.astro';
+import Posts from '$lib/components/posts/Posts.astro';
+import Page from '$lib/layouts/Page.astro';
 import Fa from 'svelte-fa';
 ---
