+++ src/pages/posts/index.astro
@@ -1,7 +1,7 @@
 ---
-import PostFilter from '$components/posts/PostFilter.svelte';
-import Posts from '$components/posts/Posts.astro';
-import Page from '$layouts/Page.astro';
+import PostFilter from '$lib/components/posts/PostFilter.svelte';
+import Posts from '$lib/components/posts/Posts.astro';
+import Page from '$lib/layouts/Page.astro';
 ---
 
 <Page>
