+++ src/pages/posts/tag/[tag].astro
@@ -1,8 +1,8 @@
 ---
-import PostFilter from '$components/posts/PostFilter.svelte';
-import Posts from '$components/posts/Posts.astro';
+import PostFilter from '$lib/components/posts/PostFilter.svelte';
+import Posts from '$lib/components/posts/Posts.astro';
 import { type Tag, tags } from '$data/tags';
-import Page from '$layouts/Page.astro';
+import Page from '$lib/layouts/Page.astro';
 
 const tag = Astro.params.tag as Tag;
