+++ src/pages/projects.astro
@@ -1,6 +1,6 @@
 ---
-import Projects from '$components/Projects.svelte';
-import Page from '$layouts/Page.astro';
+import Projects from '$lib/components/Projects.svelte';
+import Page from '$lib/layouts/Page.astro';
 ---
 
 <Page>
