+++ tsconfig.json
@@ -19,14 +19,8 @@
     "exactOptionalPropertyTypes": false,
     "baseUrl": ".",
     "paths": {
-      "$components/*": [
-        "src/components/*"
-      ],
-      "$layouts/*": [
-        "src/layouts/*"
-      ],
-      "$helpers/*": [
-        "src/helpers/*"
+      "$lib/*": [
+        "src/lib/*"
       ],
       "$data/*": [
         "src/data/*"
